Use docker buildx to build base Linux images
We use "docker buildx" when building the base Linux images for
Antrea. This gives us access to all the BuildKit capabilities, in
particular exporting the build cache to the registry as a separate
image, when using the docker-container driver. Exporting the cache means
that the base images won't need to be rebuilt from scratch every time
(whether it is for local builds or CI builds).

Using buildx with Docker requires Docker Engine 19.03 or newer, but that
version came out in 2019. Note that we already require buildx / BuildKit
since merging https://github.com/antrea-io/antrea/pull/5918.

Note that in recent Docker versions (starting with Docker Engine 23.0
and Docker Desktop 4.19), "docker build" is an alias for "docker buildx
build" (buildx is the default build client). However, explicitly using
"docker buildx build" lets us support older versions as well.

While exporting the cache to the registry requires using the
docker-container driver, and will not work with the default docker
driver, developers are not supposed to export the cache anyway. This is
reserved for Github CI jobs.

We also add a new "--no-cache" option to force re-building from
scratch. That option should not be required for normal workflows.

The Makefile is improved to reduce verbosity.

Fixes #5944

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>